+++ nokhwa-bindings-macos/src/device.rs
@@ -3,14 +3,14 @@ use crate::ffi::{
     AVCaptureWhiteBalanceGains, CGFloat, CGPoint, CMVideoFormatDescriptionGetDimensions,
     EncodableCMTime,
 };
+use crate::ffi::{
+    CMFormatDescriptionGetMediaSubType, CMFormatDescriptionRef, CMTime, CMVideoDimensions,
+};
 use crate::session::AVCaptureDeviceDiscoverySession;
 use crate::types::{AVCaptureDevicePosition, AVCaptureDeviceType, AVMediaType};
 use crate::util::{
     create_boilerplate_impl, ns_arr_to_vec, nsstr_to_str, raw_fcc_to_frameformat, str_to_nsstr,
     try_ns_arr_to_vec,
-};
-use core_media_sys::{
-    CMFormatDescriptionGetMediaSubType, CMFormatDescriptionRef, CMTime, CMVideoDimensions,
 };
 use nokhwa_core::{
     error::NokhwaError,
@@ -235,21 +235,20 @@ impl AVCaptureDevice {
                 error: "Already in use".to_string(),
             });
         }
-        let err_ptr: *mut c_void = std::ptr::null_mut();
-        let accepted: bool = unsafe { objc2::msg_send![self.inner, lockForConfiguration: err_ptr] };
-        if !err_ptr.is_null() {
-            return Err(NokhwaError::SetPropertyError {
-                property: "lockForConfiguration".to_string(),
-                value: "Locked".to_string(),
-                error: "Cannot lock for configuration".to_string(),
-            });
-        }
-        // Space these out for debug purposes
+        let mut err_ptr: *mut AnyObject = std::ptr::null_mut();
+        let accepted: bool =
+            unsafe { objc2::msg_send![self.inner, lockForConfiguration: &mut err_ptr as *mut _] };
         if !accepted {
             return Err(NokhwaError::SetPropertyError {
                 property: "lockForConfiguration".to_string(),
                 value: "Locked".to_string(),
-                error: "Lock Rejected".to_string(),
+                error: if !err_ptr.is_null() {
+                    let desc: *mut AnyObject =
+                        unsafe { objc2::msg_send![err_ptr, localizedDescription] };
+                    nsstr_to_str(desc).into_owned()
+                } else {
+                    "Lock rejected".to_string()
+                },
             });
         }
         self.locked = true;